Add the first iteration of the `util.hashing` package.
Move `MurmurHash3` to `util.hashing`.
Make the `class` private and retain a public companion
`object`, and put the `MurmurHash3.Hashing` implementations
for various types in the companion.

Add a method which composes `ByteswapHashing` with some other hashing.

Rename `hashOf` to `hash`.

Fix chi-square test in a test-case.

Review by @jsuereth.