Fix SI-5284.
The problem was the false assumption that methods specialized
on their type parameter, such as this one:

    class Foo[@spec(Int) T] {
      def bar[@spec(Int) S >: T] = ...
    }

have their normalized versions (`bar$mIc$sp`) never called
from the base specialization class `Foo`.
This meant that the implementation of `bar$mIc$sp` in `Foo`
simply threw an exception.

This assumption is not true, however. See this:

    object Baz {
      def apply[T]() = new Foo[T]
    }

Calling `Baz.apply[Int]()` will create an instance of the
base specialization class `Foo` at `Int`.
Calling `bar` on this instance will be rewritten by
specialization to calling `bar$mIc$sp`, hence the error.

So, we have to emit a valid implementation for `bar`,
obviously.
Problem is, such an implementation would have conflicting
type bounds in the base specialization class `Foo`, since
we don't know if `T` is a subtype of `S = Int`.
However, notice that the bounds are valid if and only if
`T = Int`. In the same time, `bar$mIc$sp` will only be
invoked given that the type bounds hold.
This means we can cast the expressions in method applications
to the required specialized type bound.

The following changes have been made:
1) The decision of whether or not to create a normalized
version of the specialized method is not done on the
`conflicting` relation anymore.
Instead, it's done based on the `satisfiable` relation,
which is true if there is possibly an instantiation of
the type parameters where the bounds hold.
2) The `satisfiable` method has a new variant called
`satisfiableConstraints`, which does unification to
figure out how the type parameters should be instantiated
in order to satisfy the bounds.
3) The `Duplicators` get a member called `castmap` which
maps type parameters to types that they have to be cast to.
Whenever an `Apply` tree does not have an appropriate
types for its arguments, the argument expressions are casted.

Cleanup pending.

Review by @dragos
Review by @VladUreche