Fix SI-5986.
Here we had an issue that RedBlack does not work the same way
for sets - which are not supposed to replace an element if
it is the same (wrt equals) and maps - which should replace
the corresponding values.

Adding an overwrite parameter which decides whether to overwrite
added keys if they are the same in the ordering.

Fix tests.